requirements.txt
azure-functions
azure-storage-blob
azure-identity
azure-cosmos
pyarrow
pypdf

openai
# instead of openai[embeddings] because of sklearn being in there and scipy not
scikit-learn >= 1.0.2  # Needed for embedding utils, versions >= 1.1 require python 3.8
tenacity >= 8.0.1
matplotlib
plotly
numpy
scipy
pandas >= 1.2.3  # Needed for CLI fine-tuning data preparation tool
pandas-stubs >= 1.1.0.11  # Needed for type hints for mypy
openpyxl >= 3.0.7  # Needed for CLI fine-tuning data preparation tool xlsx format
